requirements.txt
openenv-core
huggingface_hub
numpy
pandas